Remove integrations as much as we can
