+++ cmd/devices-api/web2_pair_sub_command.go
@@ -8,10 +8,8 @@ import (
 	"math/big"
 
 	"github.com/DIMO-Network/devices-api/internal/config"
-	"github.com/DIMO-Network/devices-api/internal/constants"
 	"github.com/DIMO-Network/devices-api/internal/services"
-	"github.com/DIMO-Network/devices-api/internal/services/autopi"
-	"github.com/DIMO-Network/devices-api/internal/services/macaron"
+	"github.com/DIMO-Network/devices-api/internal/services/genericad"
 	"github.com/DIMO-Network/devices-api/internal/utils"
 	"github.com/DIMO-Network/devices-api/models"
 	"github.com/DIMO-Network/shared/db"
@@ -64,34 +62,15 @@ func (p *web2PairCmd) Execute(ctx context.Context, _ *flag.FlagSet, _ ...interfa
 		p.logger.Fatal().Msgf("Can't find aftermarket device %d.", am.TokenID)
 	}
 
-	dm, err := ddSvc.GetMakeByTokenID(context.TODO(), am.DeviceManufacturerTokenID.Int(nil))
-	if err != nil {
-		p.logger.Fatal().Msgf("Can't retrieve manufacturer %d.", am.DeviceManufacturerTokenID)
-	}
-
 	p.logger.Info().Msgf("Attempting to web2 pair am device %s to vehicle %s.", amToken, vToken)
 
-	autoPiSvc := services.NewAutoPiAPIService(&p.settings, p.pdb.DBS)
 	autoPiIngest := services.NewIngestRegistrar(producer)
 	eventService := services.NewEventService(&p.logger, &p.settings, producer)
 	deviceDefinitionRegistrar := services.NewDeviceDefinitionRegistrar(producer, &p.settings)
-	hardwareTemplateService := autopi.NewHardwareTemplateService(autoPiSvc, p.pdb.DBS, &p.logger)
 
-	switch dm.Name {
-	case constants.AutoPiVendor:
-		autoPi := autopi.NewIntegration(p.pdb.DBS, ddSvc, autoPiSvc, autoPiIngest, eventService, deviceDefinitionRegistrar, hardwareTemplateService, &p.logger)
-
-		err = autoPi.Pair(ctx, amToken, vToken)
-		if err != nil {
-			p.logger.Fatal().Err(err).Msg("Pairing failure.")
-		}
-	case "Hashdog":
-		macaron := macaron.NewIntegration(p.pdb.DBS, ddSvc, autoPiIngest, eventService, deviceDefinitionRegistrar, &p.logger)
-
-		err := macaron.Pair(ctx, amToken, vToken)
-		if err != nil {
-			p.logger.Fatal().Err(err).Msg("Pairing failure.")
-		}
+	integ := genericad.NewIntegration(p.pdb.DBS, ddSvc, autoPiIngest, eventService, deviceDefinitionRegistrar, &p.logger)
+	if err := integ.Pair(ctx, amToken, vToken); err != nil {
+		p.logger.Fatal().Err(err).Msg("Pairing failure.")
 	}
 
 	p.logger.Info().Msg("Pairing success.")
